chore: refactor of auto emission script by system
